Fix for seg-fault with ScanArchive data conversion.
Replaced (ScanArchive *) parameters in getAcquisitions() member
functions in SequenceConverter and derived classes with a
ScanArchivePointer reference.  This avoids re-instantiating a
ScanArchive referrring to the same underlying object, which was then
destructed twice leading to sadness.

The corresponding change was also made for the (Pfile *) parameters
even though these were not causing seg-faults.